+++ compose/compose_par.go
@@ -1,0 +1,14 @@
+package compose
+
+import (
+	"github.com/ory/fosite"
+	"github.com/ory/fosite/handler/par"
+)
+
+// PushedAuthorizeHandlerFactory creates the basic PAR handler
+func PushedAuthorizeHandlerFactory(config fosite.Configurator, storage interface{}, strategy interface{}) interface{} {
+	return &par.PushedAuthorizeHandler{
+		Storage: storage,
+		Config:  config,
+	}
+}